Add an environment variable to skip signing.
Having to unlock a gpg key every time we run the script is pretty annoying
for debugging purposes, and theoretically someone might want to make an
unsigned repository. Either way, allow signing to be skipped by setting
SKIP_SIGN to a non-empty value; this also disabled the requirement for
KEY_ID to be set, since it's not used.