Create the temporary directory inside build_repo.
This eliminates a path parameter for that function, but also encapsulates
the temporary storage in the build functions. Whether or not a build needs
a temporary directory is not something the caller (i.e. the script itself)
needs to care about.

Also, try to be consistent about using lower-case names for internal vars,
and saving uppercase for things from the actual environment. This makes the
distinction between the two clear, and is a little bit easier to read. This
will be done elsewhere as we go along.