Separate architecture recording from package extraction.
Recording the package architecture in the set for the repo is the only part
of package info extraction that has to know where the meta directory is, so
instead just extract the architecture into a file in the package directory,
and update the repo info while we're building the package index.

Speaking of, now that we're doing that, the function isn't so much building
the package index as it is updating the repo info with the package info, so
it deserves a new name. This makes writing the package index to stdout a
bit weird now, but oh well.