Create scratch dir facility, encapsulate package meta temp dir.
In order for encapsulated temp directories to really work, a component has
to be able to get one on demand without anybody else knowing anything about
it. Previously we were doing this by passing the $TMPDIR global around
where necessary, but instead, create a facility whereby a component can
ask for a name and get one without any prior information. In theory, this
could be a freshly allocated temp directory, although in practice we create
them under the same workspace directory to make cleanup easier. Paths are
still passed where possible, because it makes the code cleaner.

This also undoes the printing of the package index to stdout. We're
updating other things, there's no reason for the Packages index to be
special. It does require an extra parameter, but that's sensible because we
need to tell the function /which/ repo to update.