Convert to archive-style repositories.
Debian only supports two types of repositories: flat (aka "trivial") and
archive-style. Flat is simple and easy, but doesn't support apt-pinning,
which is a thing we want to enable in case folks want to use it. So we're
doing archive-style repos.

The main difference is that we generates Release files for each dist (so a
dist is somewhat like a flat repository), and the package indexes are split
up by component and architecture.

The patch is too large to explain in detail, and it doesn't really admit
of being done in stages. There are a number of additional improvements that
have been made along the way, which will also have to go without comment.