Replace $KEEP_TMP with $WORK_DIR.
Rather than specifying whether to keep the temporary directory, and then
having to report it to the user (and hope they see it), allow the user to
set the work directory themselves; if they've set it, we leave it alone,
otherwise we remove it on exit. This is both convenient, and follows the
normal rules of object ownership: if we create a temp dir, we have to
delete it, but if someone hands us a directory, it's on them.