Replace the use of readlink -f to make absolute paths.
Readlink -f will resolve any symlink components when canonicalizing the
path, which may not be what the repo admin wants: they might prefer to be
able to move their package pool by updating a symlink, without having to
regenerate the repository, so we should keep the package path as-specified.

realpath(1) will do something similar with the -s argument, but I'm not
sure if that's portable, so instead just add $PWD if it's a relative path.
We don't need full canonicalization, just conversion to an absolute path,
so this is all we need.