Clear the output directory contents, rather than the directory itself.
We want to fully re-generate the repository contents, but the directory
itself may have properties we want to preserve, like special permissions.
This may also avoid problems when the repo dir has permissions to allow
the user of this script to create the repository there, but its parent does
not (the setgid bit would be a good example).